pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "SafeStreamAI"
version = "0.1.0"
description = "AI-powered content moderation environment"
authors = [
    {name = "LeahRocks"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "jinja2>=3.1.3",
    "python-multipart>=0.0.9",
    "transformers>=4.41.2",
    "torch>=2.0.0",
    "accelerate>=0.30.1",
    "numpy>=1.26.4",
    "pandas>=2.2.2",
    "scikit-learn>=1.4.2",
    "huggingface_hub>=0.23.0",
    "openai",
    "openenv-core",
    "httpx",
    "python-dotenv"
]

[project.scripts]
server = "server.app:main"

